ai_services/requirements.txt
# ai-service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
ollama
groq
openai-whisper 
pydub 
ffmpeg-python
python-multipart